feature: move downsampling to customization widget
