+++ frontend/src/renderer/pages/visualization/customizableElements/index.ts
@@ -1,0 +1,2 @@
+export * from './CustomizeTitle';
+export * from './CustomizeDownsampling';
