+++ frontend/src/renderer/types/grid.ts
@@ -4,5 +4,4 @@ export interface GridLayoutPlotProps {
   data: DataGridPlot;
   colWidth: number;
   rowHeight: number;
-  downsamplingList: string[];
 }
